% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upriver.R
\name{expand_parameters}
\alias{expand_parameters}
\title{Expand movement parameters}
\usage{
expand_parameters(parameters)
}
\arguments{
\item{parameters}{(list) Movement parameters (names: \code{rates},
\code{distances})}
}
\value{
(list) Expanded parameters
}
\description{
Expands the movement parameters to include a 0th and n+1th reach.
}
